+++ src/drive/web/modules/filelist/AddFolder.spec.jsx
@@ -16,7 +16,8 @@ jest.mock('cozy-flags', () => jest.fn())
 jest.mock('cozy-keys-lib', () => ({
   withVaultClient: jest.fn().mockReturnValue({}),
   useVaultClient: jest.fn(),
-  WebVaultClient: jest.fn().mockReturnValue({})
+  WebVaultClient: jest.fn().mockReturnValue({}),
+  withVaultUnlockContext: jest.fn().mockReturnValue({})
 }))
 describe('AddFolder', () => {
   const setup = () => {