+++ src/drive/web/modules/viewer/helpers.js
@@ -6,8 +6,8 @@ import {
 
 export const downloadFile = async (client, file, { vaultClient }) => {
   if (isEncryptedFile(file)) {
-    const encryptionKey = await getEncryptionKeyFromDirId(client, file.dir_id)
-    return downloadEncryptedFile(client, vaultClient, { file, encryptionKey })
+    const decryptionKey = await getEncryptionKeyFromDirId(client, file.dir_id)
+    return downloadEncryptedFile(client, vaultClient, { file, decryptionKey })
   } else {
     return client.collection('io.cozy.files').download(file)
   }